requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
anthropic>=0.18.0
PyPDF2>=3.0.0
jsonschema>=4.17.0
python-docx>=1.1.0

# Phase 2: PDF Generation & Visualization
reportlab>=4.0.0
matplotlib>=3.8.0
pillow>=10.0.0
plotly>=5.18.0
kaleido>=0.2.1

# Phase 3: APIs & Automation
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
schedule>=1.2.0

# Phase 4: Historical Intelligence & Advanced Analytics
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0

# Phase 5: Web Interface & Competitor Analysis
flask>=3.0.0
flask-cors>=4.0.0
